requirements.txt
discord.py
google-generativeai
dotenv